move board and game file to lib folder and kept main file in bin folder
