+++ bin/main.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
-require_relative './board.rb'
-require_relative './game.rb'
+require_relative '../lib/board.rb'
+require_relative '../lib/game.rb'
 
 new_game = Interface.new
 new_game.user_details
