+++ lib/board.rb
